Added a base implementation for server
Not all the roots are perfectly compatible with the api described
However, it is mostly a matter of rearranging urls in urls.py.
I wasn't sure how much of this to site since it was mostly from the tutorial.
Still need to add citations to README.md if we decide to keep this